refactor(app): extract stdin streaming + virtual buffers (phase 5e)
Move open_stdin_buffer, poll_stdin_streaming, complete_stdin_streaming,
is_stdin_streaming, create_virtual_buffer, set_virtual_buffer_content
out of buffer_management.rs into a new app/virtual_buffers.rs sibling.
~398 lines.

These drive the StdinStream subsystem (extracted in phase 2e) plus
helpers for creating buffers backed by virtual content. Both flows
construct a buffer + register it + populate it via specialized
content sources rather than reading from disk.

buffer_management.rs goes from 1127 -> 745 lines.
2026 tests still pass.
