+++ crates/fresh-editor/src/app/mod.rs
@@ -70,6 +70,7 @@ mod toggle_actions;
 pub mod types;
 mod undo_actions;
 mod view_actions;
+mod virtual_buffers;
 pub mod warning_domains;
 pub mod workspace;
